Refactor domain table row actions and introduce edit dialog hook
- Removed the `useDomainTableRowEditDialog` implementation from `domain-table-row-actions.tsx` and created a separate file for it.
- Added tests for the `useDomainTableRowEditDialog` hook to validate its functionality, ensuring it starts closed and toggles correctly.
- Simplified the test file for `getDomainTableRowDeleteLabel` by removing unnecessary imports and focusing on the label functionality.
